+++ e2e_playwright/multipage_apps/mpa_configure_sidebar_test.py
@@ -17,24 +17,8 @@ import os
 import pytest
 from playwright.sync_api import Page, expect
 
-from e2e_playwright.conftest import ImageCompareFunction, wait_for_app_run, wait_until
+from e2e_playwright.conftest import ImageCompareFunction, wait_for_app_run
 from e2e_playwright.shared.app_utils import get_element_by_key
-
-
-def check_logo_source_errors(messages: list[str]):
-    """Check that both logo source error messages are logged."""
-    main_logo_error_count = 0
-    sidebar_logo_error_count = 0
-
-    # iterate over messages and check if both error messages are present
-    for message in messages:
-        if "Client Error: Logo source error" in message:
-            main_logo_error_count += 1
-        if "Client Error: Sidebar Logo source error" in message:
-            sidebar_logo_error_count += 1
-
-    assert main_logo_error_count == 1
-    assert sidebar_logo_error_count == 1
 
 
 @pytest.fixture(scope="module")
@@ -131,28 +115,4 @@ def test_logo_no_sidebar(app: Page, assert_snapshot: ImageCompareFunction):
     ).to_have_attribute("href", "https://www.example.com")
     assert_snapshot(
         app.get_by_test_id("stSidebarCollapsedControl"), name="logo-no-sidebar"
-    )
-
-
-def test_logo_source_errors(app: Page, app_port: int):
-    """Test that logo source errors are logged."""
-    app.route(
-        f"http://localhost:{app_port}/media/**",
-        lambda route: route.fulfill(
-            status=404, headers={"Content-Type": "text/plain"}, body="Not Found"
-        ),
-    )
-
-    # Capture console messages
-    messages = []
-    app.on("console", lambda msg: messages.append(msg.text))
-
-    # Navigate to the app
-    app.goto(f"http://localhost:{app_port}")
-
-    # Wait until the expected error is logged, indicating CLIENT_ERROR was sent
-    # for the logo in the main app area and the sidebar
-    wait_until(
-        app,
-        lambda: check_logo_source_errors(messages),
     )
